Disable WAN check.
Periodically checks for Internet connectivity through the
check_wan_status.lua script by either pinging or resolving some fixed
set of China domains like baidu, qq, tabao and its own api.miwifi.com.

The downside of removing this is that the LED indicator on the router
will not be accurate, but I don't mind that.