Added more measures to retain router access.
Apparently setting xiaoqiang-defaults.txt is not enough. Somehow the
defaults are not getting applied after a factory reset via the reset
button. This change will always force ssh_en and uart_en to be on, and
also dropbear will always be started.

If you need to block access, either set a very long and secure password,
or use firewall rules to block the SSH port.